moved _get_pk_data to CCL
